android側同時通信完成
